bundle, verify: move shared caps to internal/limits
Move the per-bundle upper bounds package from pkg/limits/ to
internal/limits/. The constants remain uppercase since cross-package
access requires it, but the internal/ location keeps them off the
public Go API surface as requested in PR review.

The single source of truth and the parse/verify symmetry are preserved;
only the import path changes.
